Fix: Line height and small card fixes finished
