+++ src/components/cards/CardInstructions.jsx
@@ -1,14 +1,4 @@
-import styled from '@emotion/styled';
-import { BaseCard } from './Cards';
-
-export const StyledInstructionsCard = styled(BaseCard)`
-    padding: 1rem 2rem;
-
-    h3 {
-        margin-bottom: 0.5rem;
-    }
-`
-
+import { StyledInstructionsCard } from './Cards';
 
 export function InstructionsCard({ children }) {
     return (
